+++ src/plugins/replication/index.ts
@@ -159,6 +159,17 @@ export class RxReplicationStateBase<RxDocType> {
     async _run(retryOnFail = true): Promise<boolean> {
         this.runCount++;
 
+        /**
+         * The replication happens in the background anyways
+         * so we have to ensure that we do not slow down primary tasks.
+         * But not if it is the initial replication, because that might happen
+         * on the first inital loading where it is critical to get the data
+         * as fast as possible to decrease initial page load time.
+         */
+        if (this.subjects.initialReplicationComplete.getValue()) {
+            await this.collection.database.requestIdlePromise();
+        }
+
         if (this.push) {
             const ok = await this.runPush();
             if (!ok && retryOnFail) {